pyproject.toml
[project]
name = "qaagent"
version = "0.1.0"
description = "Python QA agent with MCP server and CLI tools"
readme = "README.md"
requires-python = ">=3.11,<3.13"
authors = [{ name = "Your Name" }]
license = { text = "Proprietary" }
keywords = ["qa", "testing", "agent", "mcp", "automation"]

dependencies = [
    "typer>=0.12",
    "pydantic>=2.7",
    "rich>=13.7",
    "pytest>=7.4",
    "PyYAML>=6.0",
    "jinja2>=3.1",
    "faker>=20.0",
]

[project.optional-dependencies]
mcp = ["mcp"]
api = ["schemathesis", "httpx>=0.27", "PyYAML>=6.0"]
ui = [
    "playwright",
    "pytest-playwright>=0.5",
    "pytest-html>=4.1",
]
report = ["jinja2>=3.1"]
config = ["python-dotenv>=1.0"]
llm = [
    "ollama>=0.1.0",
    "litellm>=1.40.0; python_version >= '3.9'",
]
cov = ["pytest-cov>=4.1"]
perf = ["locust"]
dev = [
    "pytest-asyncio>=0.23",
]

[project.scripts]
qaagent = "qaagent.cli:main"
qaagent-mcp = "qaagent.mcp_server:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"qaagent" = [
    "templates/*.j2",
    "templates/**/*.j2",
]
